🚧 ✨ 🔥 Old tradfri -> to new Smart Devices
