+++ homecontrol/__main__.py
@@ -1,10 +1,8 @@
-from .tradfri.tradfri_gateway import TradfriGateway
-from .tradfri.motion_sensor import MotionSensor
-from .network import Network
-from .weather import Weather
+from .data.network import Network
+from .data.weather import Weather
 from .controller import Controller
 from .webapi import flask_api
-from .thread import start_thread
+from .utils.thread import start_thread
 from apscheduler.schedulers.background import BackgroundScheduler
 
 
@@ -20,9 +18,7 @@ def main():
     scheduler.start()
 
     start_thread(Network.update, seconds_between_calls=5)
-    start_thread(TradfriGateway.update, seconds_between_calls=3600 * 24)
     start_thread(Controller.update_all, seconds_between_calls=1, delay=10)
-    # start_thread(MotionSensor.update_all, seconds_between_calls=10) NOTE doesn't work yet
 
     # Web API
     flask_api.run_api()
