+++ homecontrol/smart_interfaces/effects/effect.py
