+++ homecontrol/smart_interfaces/effects/sunrise_bamboo.py
