+++ homecontrol/webapi/power.py
@@ -1,7 +1,6 @@
 from typing import Any, Dict
-from . import success, execute
-from ..tradfri import get_light_and_groups
-from ..tradfri.tradfri_gateway import TradfriGateway
+from . import success, execute, trim_name
+from ..smart_interfaces import SmartInterfaces
 from flask import Blueprint, request, abort
 
 power_blueprint = Blueprint("power", __package__)
@@ -18,23 +17,25 @@ def power() -> str:
     if not "name" in body:
         abort(400, 'Missing "name" in body')
 
-    lights_and_groups = get_light_and_groups(body["name"])
+    name = trim_name(body["name"])
+    interfaces = SmartInterfaces.get_interfaces(name)
 
     if isinstance(body["value"], str):
         body["value"] = str(body["value"]).lower()
 
-    if body["value"] == "on" or body["value"] == 1:
-        action = TradfriGateway.turn_on
-    elif body["value"] == "off" or body["value"] == 0:
-        action = TradfriGateway.turn_off
-    elif body["value"] == "toggle":
-        action = TradfriGateway.toggle
-    else:
-        abort(
-            400,
-            'field "value" has invalid value. Valid values are: "on"/"off"/1/0/"toggle"',
-        )
+    for interface in interfaces:
+        if body["value"] == "on" or body["value"] == 1:
+            action = interface.turn_on
+        elif body["value"] == "off" or body["value"] == 0:
+            action = interface.turn_off
+        elif body["value"] == "toggle":
+            action = interface.toggle
+        else:
+            abort(
+                400,
+                'field "value" has invalid value. Valid values are: "on"/"off"/1/0/"toggle"',
+            )
 
-    execute(body, action, args=[lights_and_groups])
+        execute(body, action)
 
     return success()
